Move cleaned name reviewing out of helpers for circular dep
