+++ zavod/zavod/helpers/names.py
@@ -6,22 +6,8 @@ from normality import squash_spaces
 from rigour.data.names.data import STOPPHRASES
 from rigour.names.check import is_nullword
 
-from zavod import settings
 from zavod.context import Context
 from zavod.entity import Entity
-from zavod.extract.names.clean import (
-    LLM_MODEL_VERSION,
-    CleanNames,
-    RawNames,
-)
-
-# alias clean_names so that it could be imported from here
-from zavod.extract.names.clean import clean_names as clean_names
-from zavod.stateful.review import (
-    JSONSourceValue,
-    Review,
-    review_extraction,
-)
 
 REGEX_AND = re.compile(r"(\band\b|&|\+)", re.I)
 REGEX_LNAME_FNAME = re.compile(r"^\w+, \w+$", re.I)
@@ -280,107 +266,3 @@ def is_name_irregular(entity: Entity, string: Optional[str]) -> bool:
         return True
 
     return False
-
-
-def apply_names(
-    entity: Entity,
-    strings: List[Optional[str]],
-    review: Review[CleanNames],
-    alias: bool = False,
-    lang: Optional[str] = None,
-) -> None:
-    field_props = [
-        ("full_name", "alias" if alias else "name"),
-        ("alias", "alias"),
-        ("weak_alias", "weakAlias"),
-        ("previous_name", "previousName"),
-    ]
-    if not review.accepted:
-        prop = "alias" if alias else "name"
-        for string in strings:
-            entity.add(prop, string, lang=lang)
-        return
-
-    for field_name, prop in field_props:
-        for name in getattr(review.extracted_data, field_name):
-            entity.add(
-                prop,
-                name,
-                lang=lang,
-                origin=review.origin,
-                original_value=review.source_value,
-            )
-
-
-def review_names(
-    context: Context,
-    entity: Entity,
-    strings: List[Optional[str]],
-    lang: Optional[str] = None,
-    alias: bool = False,
-) -> Optional[Review[CleanNames]]:
-    """
-    Clean names if needed, then post them for review.
-
-    Args:
-        context: The current context.
-        entity: The entity to apply names to.
-        string: The raw name(s) string.
-        alias: If this is known to be an alias and not a primary name.
-        lang: The language of the name, if known.
-    """
-    strings = [s for s in strings if s]
-
-    if not strings:
-        return None
-
-    if settings.CI or not any(is_name_irregular(entity, s) for s in strings):
-        prop = "alias" if alias else "name"
-        for string in strings:
-            entity.add(prop, string, lang=lang)
-        return None
-
-    non_blank_strings = [s for s in strings if s and s.strip()]
-    raw_names = RawNames(entity_schema=entity.schema.name, strings=non_blank_strings)
-    names = clean_names(context, raw_names)
-
-    source_value = JSONSourceValue(
-        key_parts=[entity.schema.name] + non_blank_strings,
-        label="names",
-        data=raw_names.model_dump(),
-    )
-    review = review_extraction(
-        context,
-        source_value=source_value,
-        original_extraction=names,
-        origin=LLM_MODEL_VERSION,
-    )
-    return review
-
-
-def apply_reviewed_names(
-    context: Context,
-    entity: Entity,
-    strings: List[Optional[str]],
-    lang: Optional[str] = None,
-    alias: bool = False,
-) -> None:
-    """
-    Clean names if needed, then post them for review.
-    Cleaned names are applied to an entity if accepted, falling back
-    to applying the original string as the name or alias if not.
-
-    Also falls back to applying the original string if the CI environment
-    variable is truthy, so that crawlers using this can run in CI.
-
-    Args:
-        context: The current context.
-        entity: The entity to apply names to.
-        string: The raw name(s) string.
-        alias: If this is known to be an alias and not a primary name.
-        lang: The language of the name, if known.
-    """
-    review = review_names(context, entity, strings, lang=lang, alias=alias)
-    if review is None:
-        return
-    apply_names(entity, strings, review, alias=alias, lang=lang)
